TaskController

+++ app/Http/Controllers/TasksController.php
@@ -76,10 +76,13 @@ class TasksController extends Controller
         $task = Task::find($id);
         // $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
         // $count_tasks = $user->tasks()->count();
-
+        
+        if (\Auth::user()->id === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -112,10 +115,13 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         $task = Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id) {
         $task->content = $request->content;
         $task->status = $request->status; 
         $task->save();
-
+        }
+        
         return redirect('/');
        
     }
@@ -129,7 +135,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id) {
         $task->delete();
+        }
 
         return redirect('/');
     }
